third_party/blink/web_tests/platform/win/external/wpt/mathml/presentation-markup/operators/stretchy-largeop-with-default-font-2-expected.txt

This is a testharness.js-based test.
[FAIL] Test that non-stretchy largeops are bigger in display mode.
  assert_greater_than: width of largeop '⨀' expected a number greater than 14 but got 14
Harness: the test ran to completion.
